+++ app/Http/Controllers/Admin/NavigationController.php
@@ -25,49 +25,6 @@ class NavigationController extends LaravelResourceController
         ]);
     }
 
-    public function updateAll(Request $request)
-    {
-        $data = $request->all();
-
-        $flattenedData = [];
-
-        foreach ($data['navigation'] as $key => $value) {
-            $flattenedData[] = $value;
-
-            $children = isset($value['children']) ? $value['children'] : null;
-
-            if ($children) {
-                foreach ($children as $key => $child) {
-                    $child['order'] = $key;
-                    $child['extra_attributes'] = [
-                        'description' => $child['description'] ?? null,
-                        'image' => $child['image'] ?? null,
-                        'type' => $child['type'] ?? null,
-                        'icon' => $child['icon'] ?? null,
-                        'column' => $child['column'] ?? null,
-                    ];
-                    $flattenedData[] = $child;
-                }
-            }
-
-        }
-
-        foreach ($flattenedData as $key => $value) {
-            $navigation = Navigation::updateOrCreate(['id' => $value['id']],
-                [
-                    'parent_id' => $value['parent_id'],
-                    'name' => $value['name'],
-                    'lang' => $value['lang'],
-                    'url' => $value['url'],
-                    'order' => $value['order'],
-                    'is_active' => $value['is_active'],
-                    'extra_attributes' => $value['extra_attributes'],
-                ]);
-        }
-
-        return back();
-    }
-
     /**
      * Show the form for creating a new resource.
      *
