+++ resources/js/Components/Public/Nav/MainMenu.vue
@@ -1,19 +1,38 @@
 <template>
-  <NMenu
-    v-model:value="activeMenuKey"
-    mode="vertical"
-    @update:value="$emit('close:drawer')"
-  />
+  <NavigationMenu v-model="modelValue" as="div">
+    <div class="mr-8">
+      <slot name="additional" />
+    </div>
+    <NavigationMenuList>
+      <NavigationMenuItem v-for="item in mainNavigation" :key="item.name">
+        <NavigationMenuTrigger
+          class="bg-transparent px-2 py-1.5 hover:bg-zinc-100 dark:bg-transparent dark:hover:bg-zinc-700 max-lg:text-xs">
+          {{ item.name }}
+        </NavigationMenuTrigger>
+        <MainNavigationMenuContent :item :for-admin-edit="false" @close-menu="closeMenu" />      
+      </NavigationMenuItem>
+    </NavigationMenuList>
+  </NavigationMenu>
 </template>
 
-<script setup lang="tsx">
-import { NMenu } from "naive-ui";
-import { ref } from "vue";
-import { usePage } from "@inertiajs/vue3";
+<script setup lang="ts">
+import {
+  NavigationMenu,
+  NavigationMenuItem,
+  NavigationMenuList,
+  NavigationMenuTrigger,
+} from '@/Components/ShadcnVue/ui/navigation-menu'
+import { computed, ref } from 'vue';
+import { usePage } from '@inertiajs/vue3';
 
-defineEmits<{
-  (e: "close:drawer"): void;
-}>();
+import MainNavigationMenuContent from './MainNavigationMenuContent.vue';
 
-const activeMenuKey = ref(usePage().props.navigationItemId);
+const modelValue = ref(null);
+
+const mainNavigation = computed(() => usePage().props.mainNavigation);
+
+function closeMenu(event: MouseEvent) {
+  //event.preventDefault();
+  modelValue.value = null;
+}
 </script>
